Make BinHandler run environment inheritance optional
Currently the environment of the test process is inherited to the test
binary, so every environment where the tests are run may have different
results if the test binary depends on env, which makes the test runs
less deterministic.

Add an option to disable inheriting the environment so that the tests
can be always run with the same env variables.

Leave the inheritance on by default to not break existing systems.